==> notebooks/01_data_mining/01. pobranie projektów.ipynb <==
import pandas as pd

## Oczyszczenie danych projektów

projects = pd.read_pickle('./../../data/00_db/projects.pkl')

### Informacje o zbiorze

projects.shape

projects.dtypes

projects.describe(include=['object'])

projects.info()

projects.info(memory_usage='deep')

projects.memory_usage(deep=True)

projects.head()

### Usuwanie zbędnych danych

projects[projects.deleted != 0].shape

Skoro wszystkie wiersze mają tę samą wartość w kolumnie `deleted` (ustawioną na `0`), można bez straty informacji tę kolumnę usunąć. Usunwam również inne kolumny, nie mające znaczenia dla analizy.

projects.drop(['deleted', 'ext_ref_id', 'url', 'owner_id', 'description', 'forked_from'], axis=1, inplace=True)

projects.head()

projects.shape

Usunięcie wierszy z brakującymi danymi, na podstawie których nie będzie można prowadzić sensownych analiz

projects.isnull().sum()

projects.dropna(subset=['language'], how='any', inplace=True)

projects.shape

projects.isnull().sum()

### Zmiana typów
Domyślnie dane przechowywane napisów przechowywane są jako `object`. Jednak, jeśli jest ograniczona ilość napisów, tak jak w przypadku kolumny `language`, można zmienić typ na `category` optymalizując pamięć.

projects.dtypes

projects.memory_usage(deep=True)

projects.language = projects.language.astype('category')
projects.created_at = projects.created_at.astype('datetime64[ns]')

projects.dtypes

Zmiana nazw kolumn

projects.rename(columns={'id': 'project_id'}, inplace=True)
projects.head()

### Eksport do pliku `.pkl` wstępnie oczyszczonych danych projektów
> UWAGA! lepiej wczytywać dane do `.pkl`, a nie do `.csv`, ponieważ w tym wypadku zapisujemy całą strukturę, nie tracimy informacji o typach i nie mamy problemu z indeksami. Sam plik zajmuje również zdecydowanie mniej pamięci.

projects.to_pickle('../../data/01_data_from_db/projects.pkl')
