==> notebooks/01_data_mining/02. pobranie commitów projektów.ipynb <==
import pandas as pd

## Oczyszczenie danych commitów

commits = pd.read_pickle('./../../data/00_db/commits.pkl')

### Informacje o zbiorze

commits.shape

commits.dtypes

commits.info()

commits.info(memory_usage='deep')

commits.memory_usage(deep=True)

commits.head()

### Usuwanie zbędnych danych

commits.drop(['sha', 'author_id', 'ext_ref_id'], axis=1, inplace=True)

commits.head()

Usunięcie wierszy z brakującymi danymi, na podstawie których nie będzie można prowadzić sensownych analiz

commits.isnull().sum()

Commity nie zawierają żadnych brakujących danych, więc nie wymagają oczyszczania

## Zmiana typów

commits.dtypes

Żaden typ nie wymaga zmiany

## Filtrowanie

Filtrowanie tylko tych commitów, które są stowarzyszone z jakimiś ID śledzonych projektów

projects = pd.read_pickle('../../data/01_data_from_db/projects.pkl')
projects.head()

commits = commits[commits.project_id.isin(projects.project_id)]

### Ponowne sprawdzenie informacji o zbiorze

commits.shape

commits.info(memory_usage='deep')

Zmiana nazw kolumn

commits.rename(columns={'id': 'commit_id'}, inplace=True)
commits.head()

### Eksport do `.pkl`

commits.to_pickle('../../data/01_data_from_db/commits.pkl')
