==> notebooks/01_data_mining/03. pobranie komentarzy commitów.ipynb <==
import pandas as pd

## Oczyszczenie danych komentarzy commitów

commits_comments = pd.read_pickle('./../../data/00_db/commit_comments.pkl')

### Informacje o zbiorze

commits_comments.shape

commits_comments.dtypes

commits_comments.info()

commits_comments.info(memory_usage='deep')

commits_comments.head()

### Usuwanie zbędnych danych

`user_id` oraz `ext_ref_id` nie są potrzebne, tak samo jak `line` i `position`, które dotyczą konkretnych fragmentów, których dotyczy komentarz  

**nie do końca wiem czym jest `comment_id`, skoro mam `id`, wydaje sie to dziwnym duplikatem, przy czym wartości są różne w obu kolumnach**

commits_comments.drop(['user_id', 'line', 'position', 'ext_ref_id', 'comment_id'], axis=1, inplace=True)

commits_comments.shape

commits_comments.isnull().sum()

Nie ma żadnych brakującyc danych, komentarze nie wymagają więc oczyszczenia

### Zmiana typów

commits_comments.dtypes

Żaden typ nie wymaga zmiany

### Filtrowanie

Filtrowanie tylko tych komentarzy, które dotyczą commitów, które mam

commits = pd.read_pickle('../../data/01_data_from_db/commits.pkl')
commits.head()

commits_comments = commits_comments[commits_comments.commit_id.isin(commits.commit_id)]

Ponowne sprawdzenie informacji o zbiorze

commits_comments.shape

commits_comments.info(memory_usage='deep')

Zmiana nazw kolumn

commits_comments.rename(columns={'id': 'commit_comment_id'}, inplace=True)
commits_comments.head()

### Eksport do `.pkl`

commits_comments.to_pickle('../../data/01_data_from_db/commit_comments.pkl')
