==> notebooks/01_data_mining/04. Pobranie issues projektów.ipynb <==
import pandas as pd

## Oczyszczenie danych issues

issues = pd.read_pickle('./../../data/00_db/issues.pkl')

### Informacje o zbiorze

issues.shape

issues.dtypes

issues.info()

issues.info(memory_usage='deep')

issues.head()

### Usuwanie zbędnych danych

issues.drop(['reporter_id', 'assignee_id', 'issue_id', 'pull_request', 'pull_request_id', 'ext_ref_id'], axis=1, inplace=True)

issues.shape

Nie ma żadnych brakujących danych

### Zmiana typów

issues.dtypes

issues.memory_usage(deep=True)

Ponieważ jeden wiersz zawiera niepoprawnie sformatowaną datę, muszę użyć innej funkcji do konwersji, niezawierającej argumentu `inplace`

issues.created_at = pd.to_datetime(issues.created_at, errors='coerce')

issues.dtypes

### Usunięcie `NaN`

issues.isnull().sum()

issues.dropna(subset=['created_at'], how='any', inplace=True)

issues.shape

issues.isnull().sum()

### Filtrowanie

Filtrowanie tylko tych issues, które są przypisane do projektów, które analizuję

projects = pd.read_pickle('../../data/01_data_from_db/projects.pkl')
projects.head()

issues = issues[issues.repo_id.isin(projects.project_id)]

### Ponowne sprawdzenie informacji o zbiorze

issues.shape

issues.info(memory_usage='deep')

issues.rename(columns={'id': 'issue_id', 'repo_id': 'project_id'}, inplace=True)
issues.head()

### Eksport do '.pkl'

issues.to_pickle('../../data/01_data_from_db/issues.pkl')
