==> notebooks/01_data_mining/05. pobranie komentarzy issues.ipynb <==
import pandas as pd

## Oczyszczenie danych komentarzy issues

issue_comments = pd.read_pickle('./../../data/00_db/issue_comments.pkl')

### Informacje o zbiorze

issue_comments.shape

issue_comments.dtypes

issue_comments.info()

issue_comments.info(memory_usage='deep')

issue_comments.head()

### Usuwanie zbędnych danych

issue_comments.drop(['user_id', 'ext_ref_id'], axis=1, inplace=True)

issue_comments.shape

issue_comments.isnull().sum()

Nie ma żadnych brakujących danych, komentarze nie wymagają więc oczyszczenia

### Zmiana typów

issue_comments.dtypes

issue_comments.comment_id = issue_comments.comment_id.astype('int64')

issue_comments.dtypes

### Filtrowanie

Filtrowanie tylko tych komentarzy, które dotyczą issues, które analizuję

issues = pd.read_pickle('../../data/01_data_from_db/issues.pkl')
issues.head()

issue_comments = issue_comments[issue_comments.issue_id.isin(issues.issue_id)]

Ponowne sprawdzenie informacji o zbiorze

issue_comments.shape

issue_comments.info(memory_usage='deep')

issue_comments.head()

### Eksport do `.pkl`

issue_comments.to_pickle('../../data/01_data_from_db/issue_comments.pkl')
