==> notebooks/01_data_mining/06. pobranie pull requestów.ipynb <==
import pandas as pd

## Pobranie i oczyszczenie danych pull requestów

pull_requests = pd.read_pickle('./../../data/00_db/pull_requests.pkl')

### Informacje o zbiorze

pull_requests.shape

pull_requests.dtypes

pull_requests.info()

pull_requests.info(memory_usage='deep')

pull_requests.head()

### Usuwanie zbędnych danych 

Nie będę korzystał z danych `head_repo` oraz `head_commit` - interesuje mnie tylko, z którego projektu był robiony fork, a to jest równoważne z projektem podanym w `base_repo_id`.

Nie jest natomiast dla mnie do końca jasne co kryje się pod `pullreq_id`, jednak stosowanie tego pola przy łączeniu z komentarzami daje bardzo dużo braków, a `id` nie daje żadnych - na tej podstawie wnioskuję, żeby korzystać z głównego `id` w przypadku tej tabeli, a drugą usuwam, aby nie zaciemniać obrazu danych.

pull_requests.isnull().sum()

pull_requests.drop(['head_repo_id', 'head_commit_id', 'base_commit_id', 'user_id', 'intra_branch', 'pullreq_id'], axis=1, inplace=True)

### Zmiana typów

pull_requests.dtypes

Żaden typ nie wymaga zmiany

### Filtrowanie

Filtruję po polu `base_repo_id` w efekcie otrzymuję więcej wierszy spełniających warunki, niż jakbym filtrował po `head_repo_id`.

projects = pd.read_pickle('../../data/01_data_from_db/projects.pkl')
projects.head()

pull_requests = pull_requests[pull_requests.base_repo_id.isin(projects.project_id)]

Ponowne sprawdzenie informacji o zbiorze

pull_requests.shape

pull_requests.info(memory_usage='deep')

Zmiana nazw kolumn

pull_requests.rename(columns={'id': 'pull_request_id', 'base_repo_id': 'project_id'}, inplace=True)
pull_requests.head()

### Eksport do `.pkl`

pull_requests.to_pickle('../../data/01_data_from_db/pull_requests.pkl')
