==> notebooks/01_data_mining/07. pobranie historii pull requestów.ipynb <==
import pandas as pd

## Oczyszczenie danych historii pull requestów

pr_history = pd.read_pickle('./../../data/00_db/pull_request_history.pkl')

### Informacje o zbiorze

pr_history.shape

pr_history.dtypes

pr_history.info()

pr_history.info(memory_usage='deep')

pr_history.head()

### Usuwanie zbędnych danych

pr_history.drop(['ext_ref_id', 'actor_id'], axis=1, inplace=True)

pr_history.shape

pr_history.isnull().sum()

### Zmiana typów

pr_history.dtypes

pr_history.memory_usage(deep=True)

pr_history.action = pr_history.action.astype('category')

pr_history.memory_usage(deep=True)

### Filtrowanie

Filtruję tylko te dane historii, które dotyczą analizowanych przeze mnie pull requestów.

pull_requests = pd.read_pickle('../../data/01_data_from_db/pull_requests.pkl')
pull_requests.head()

pr_history.shape

pr_history = pr_history[pr_history.pull_request_id.isin(pull_requests.pull_request_id)]

pr_history.isnull().sum()

Ponowne sprawdzenie informacji o zbiorze

pr_history.shape

pr_history.info(memory_usage='deep')

pr_history.rename(columns={'id': 'pull_request_history_id'}, inplace=True)
pr_history.head()

### Eksport do `.pkl`

pr_history.to_pickle('../../data/01_data_from_db/pull_request_history.pkl')
