==> notebooks/01_data_mining/08. pobranie komentarzy pull requestów.ipynb <==
import pandas as pd

## Pobranie i oczyszczenie danych komentarzy commitów

pr_comments = pd.read_pickle('./../../data/00_db/pull_request_comments.pkl')

### Informacje o zbiorze

pr_comments.shape

pr_comments.dtypes

pr_comments.info()

pr_comments.info(memory_usage='deep')

pr_comments.head()

### Usuwanie zbędnych danych

pr_comments.drop(['user_id', 'position', 'commit_id', 'ext_ref_id'], axis=1, inplace=True)

pr_comments.shape

pr_comments.isnull().sum()

### Zmiana typów

pr_comments.dtypes

pr_comments.comment_id = pr_comments.comment_id.astype('int64')

pr_comments.dtypes

pr_comments.isnull().sum()

### Filtrowanie

Filtrowanie tylko tych komentarzy, które dotyczą analizowanych pull requestów

pull_requests = pd.read_pickle('../../data/01_data_from_db/pull_requests.pkl')
pull_requests.head()

pr_comments.shape

pr_comments[pr_comments.pull_request_id.isin(pull_requests.pull_request_id)].shape

pr_comments = pr_comments[pr_comments.pull_request_id.isin(pull_requests.pull_request_id)]

Ponowne sprawdzenie informacji o zbiorze

pr_comments.info(memory_usage='deep')

pr_comments.head()

### Eksport do `.pkl`

pr_comments.to_pickle('../../data/01_data_from_db/pull_request_comments.pkl')
