==> notebooks/01_data_mining/09. Połączenie pull requestów z historią.ipynb <==
import pandas as pd

pull_requests = pd.read_pickle('../../data/01_data_from_db/pull_requests.pkl')
pull_requests.head()

pr_history = pd.read_pickle('../../data/01_data_from_db/pull_request_history.pkl')
pr_history.head()

Merge kolumn przy użyciu kolumny `pull_request_id`

pr_with_history = pd.merge(pull_requests, pr_history, on=['pull_request_id'], how='left', sort=False)
pr_with_history.head()

Usunięcie braków

pr_with_history.isnull().sum()

pr_with_history.shape

pr_with_history.dropna(how='any', inplace=True)

pr_with_history.shape

Konwersja typów

pr_with_history.dtypes

pr_with_history.pull_request_history_id = pr_with_history.pull_request_history_id.astype('int64')

pr_with_history.dtypes

pr_with_history.head()

### Eksport do `.pkl`

pr_with_history.to_pickle('../../data/01_data_from_db/pull_requests_with_history.pkl')
