==> notebooks/01_data_mining/10. pobranie obserwujących (gwiazdkujących).ipynb <==
import pandas as pd

## Pobranie i oczyszczenie danych obserwujących

watchers = pd.read_pickle('./../../data/00_db/watchers.pkl')

### Informacje o zbiorze

watchers.shape

watchers.dtypes

watchers.info()

watchers.info(memory_usage='deep')

watchers.head()

### Usuwanie zbędnych danych

watchers.drop(['user_id', 'ext_ref_id'], axis=1, inplace=True)

watchers.shape

watchers.isnull().sum()

Nie ma żadnych brakujących danych

### Zmiana typów

watchers.dtypes

Żadna kolumna nie wymaga zmiany typu

### Filtrowanie

Filtrowanie tylko tych obserwacji, które dotyczą analizowanych projektów

projects = pd.read_pickle('../../data/01_data_from_db/projects.pkl')
projects.head()

watchers = watchers[watchers.repo_id.isin(projects.project_id)]

Ponowne sprawdzenie informacji o zbiorze

watchers.shape

watchers.info(memory_usage='deep')

watchers.rename(columns={'repo_id': 'project_id'}, inplace=True)
watchers.head()

### Eksport do `.pkl`

watchers.to_pickle('../../data/01_data_from_db/watchers.pkl')
